requirements.txt
langchain
python-dotenv
langchain_community
pypdf
beautifulsoup4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
 langchain-community
 sentence_transformers
 langchain_huggingface
 faiss-cpu
 langchain-chroma
langchain-groq
langchain_core
fastapi
uvicorn
langserve
supabase
sounddevice==0.4.6
soundfile==0.12.1
numpy>=1.24.3
openai>=1.3.0